Change naming and remove unnecessary try catch
